Indicates that the branch is working on the business layer of the layered architecture.
