+++ request.rest
@@ -5,25 +5,25 @@ POST http://localhost:3003/users
 Content-Type: application/json
 
   {
-    "id": "u005",
-    "name": "André",
-    "email": "andre@email.com",
-    "password": "80239212",
+    "id": "u004",
+    "name": "Amandinha",
+    "email": "amandinha@email.com",
+    "password": "44239212",
     "role": "user"
   }
 ###
 
-PUT http://localhost:3003/users/u006
+PUT http://localhost:3003/users/u009
 Content-Type: application/json
 
   {
     "id": "u006",
-    "name": "Beltrano",
-    "email": "beltrano@email.com",
-    "password": "90239111",
+    "name": "Joãozinho",
+    "email": "joazinho@email.com",
+    "password": "88239111",
     "role": "admin"
   }
 ###
 
-delete http://localhost:3003/users/u006
+delete http://localhost:3003/users/u004
 ###
