+++ src/controller/UserController.ts
@@ -1,28 +1,17 @@
 import { Request, Response } from "express";
-import { UserDatabase } from "../database/UserDatabase";
-import { User } from "../models/users";
-import { UserDB } from "../types";
+import { UserBusiness } from "../business/UserBusiness";
 
 export class UserController {
   public getUsers = async (req: Request, res: Response) => {
     try {
-      const q = req.query.q as string | undefined;
+      const input = {
+        q: req.query.q as string | undefined,
+      };
 
-      const userDatabase = new UserDatabase();
-      const usersDB = await userDatabase.findUsers(q);
+      const userBusiness = new UserBusiness();
+      const response = await userBusiness.getUsers(input);
 
-      const users: User[] = usersDB.map(
-        (userDB) =>
-          new User(
-            userDB.id,
-            userDB.name,
-            userDB.email,
-            userDB.password,
-            userDB.role,
-            userDB.created_at
-          )
-      );
-      res.status(200).send(users);
+      res.status(200).send(response);
     } catch (error) {
       console.log(error);
 
@@ -40,68 +29,20 @@ export class UserController {
 
   public createUsers = async (req: Request, res: Response) => {
     try {
-      const { id, name, email, password, role } = req.body;
-
-      if (typeof id !== "string" || id.length < 4) {
-        res.statusCode = 404;
-        throw new Error(
-          "O campo 'id' deve ser uma string com pelo menos 4 caracteres"
-        );
-      }
-
-      if (typeof name !== "string" || name.length < 3) {
-        res.statusCode = 404;
-        throw new Error(
-          "O campo 'nome' deve ser uma string com pelo menos 3 caracteres"
-        );
-      }
-
-      if (!email || !email.includes("@")) {
-        res.statusCode = 404;
-        throw new Error(
-          `O campo 'email' deve ser um endereço de e-mail válido`
-        );
-      }
-
-      if (typeof password !== "string" || password.length < 6) {
-        res.statusCode = 404;
-        throw new Error(`O campo 'password' deve ter pelo menos 6 caracteres.`);
-      }
-      if (typeof role !== "string" || role.length < 4) {
-        res.statusCode = 404;
-        throw new Error(
-          `O campo 'role' deve ser uma string com pelo menos 4 caracteres`
-        );
-      }
-
-      const userDatabase = new UserDatabase();
-      const userDBExists = await userDatabase.findUserById(id);
-
-      if (userDBExists) {
-        res.status(400);
-        throw new Error("'id' já existente");
-      }
-
-      const newUser = new User(
-        id,
-        name,
-        email,
-        password,
-        role,
-        new Date().toISOString()
-      );
-
-      const newUserDB: UserDB = {
-        id: newUser.getId(),
-        name: newUser.getName(),
-        email: newUser.getEmail(),
-        password: newUser.getPassword(),
-        role: newUser.getRole(),
-        created_at: newUser.getCreatedAt(),
+      const input = {
+        id: req.body.id,
+        name: req.body.name,
+        email: req.body.email,
+        password: req.body.password,
+        role: req.body.role,
       };
 
-      await userDatabase.insertUser(newUserDB);
-      res.status(201).send(newUser);
+      const userBusiness = new UserBusiness();
+      const response = await userBusiness.createUsers(input);
+
+      res
+        .status(201)
+        .send(`Usuário ${response.getName()} criado com sucesso!!`);
     } catch (error) {
       console.log(error);
 
@@ -119,56 +60,18 @@ export class UserController {
 
   public updateUsers = async (req: Request, res: Response) => {
     try {
-      const { id } = req.params;
-      const { name, email, password, role } = req.body;
+      const input = {
+        id: req.params.id,
+        name: req.body.name,
+        email: req.body.email,
+        password: req.body.password,
+        role: req.body.role,
+      };
 
-      if (typeof id !== "string" || id.length < 4) {
-        res.statusCode = 404;
-        throw new Error(
-          "O campo 'id' deve ser uma string com pelo menos 4 caracteres"
-        );
-      }
+      const userBusiness = new UserBusiness();
+      const response = await userBusiness.updateUsers(input);
 
-      if (typeof name !== "string" || name.length < 3) {
-        res.statusCode = 404;
-        throw new Error(
-          "O campo 'nome' deve ser uma string com pelo menos 3 caracteres"
-        );
-      }
-
-      if (!email || !email.includes("@")) {
-        res.statusCode = 404;
-        throw new Error(
-          `O campo 'email' deve ser um endereço de e-mail válido`
-        );
-      }
-
-      if (typeof password !== "string" || password.length < 6) {
-        res.statusCode = 404;
-        throw new Error(`O campo 'password' deve ter pelo menos 6 caracteres.`);
-      }
-      if (typeof role !== "string" || role.length < 4) {
-        res.statusCode = 404;
-        throw new Error(
-          `O campo 'role' deve ser uma string com pelo menos 4 caracteres`
-        );
-      }
-
-      const userDatabase = new UserDatabase();
-      const userDBExists = await userDatabase.findUserById(id);
-
-      if (!userDBExists) {
-        res.status(400);
-        throw new Error("Usuário não econtrado");
-      }
-
-      userDBExists.name = name;
-      userDBExists.email = email;
-      userDBExists.password = password;
-      userDBExists.role = role;
-
-      await userDatabase.updateUser(userDBExists);
-      res.status(201).send("Usuário atualizado com sucesso");
+      res.status(201).send(`Cadastro atualizado com sucesso ${response.name}`);
     } catch (error) {
       console.log(error);
 
@@ -186,23 +89,14 @@ export class UserController {
 
   public deleteUsers = async (req: Request, res: Response) => {
     try {
-      const { id } = req.params;
+      const input = {
+        id: req.params.id,
+      };
 
-      if (typeof id !== "string") {
-        res.statusCode = 404;
-        throw new Error("O campo 'id' deve ser umas string");
-      }
+      const userBusiness = new UserBusiness();
+      const response = await userBusiness.deleteUsers(input);
 
-      const userDatabase = new UserDatabase();
-      const userDBExists = await userDatabase.findUserById(id);
-
-      if (!userDBExists) {
-        res.statusCode = 404;
-        throw new Error("Não foi possível encontrar o usuário");
-      }
-
-      await userDatabase.deleteUser(id);
-      res.status(201).send("Usuário deletado com sucesso");
+      res.status(200).send(`Usuário ${response.name} deletado com sucesso!!`);
     } catch (error) {
       console.log(error);
